portfolio/demo-fastapi-llm/requirements.txt
fastapi
uvicorn[standard]
openai
pydantic-settings
python-dotenv
pytest
